#137: extract and generalize modal
