+++ src/components/conference/NewConference/Page/PerformanceModal/NewPerformanceModal.svelte
@@ -1,0 +1,10 @@
+<script>
+    import Modal from "../../Modal.svelte";
+</script>
+
+
+<Modal heading="Legg til bidrag">
+    <h2></h2>
+
+    <!--Form /-->
+</Modal>
