+++ src/routes/konferanser/+page.svelte
@@ -8,7 +8,7 @@
 	import NewConferenceModal from '../../components/conference/NewConference/Modal/NewConferenceModal.svelte';
 	import type { IConferencesPageLoadData } from './+page.server';
 	import type { ConferenceCategoryType } from '../../enums/conference-category';
-    import {displayNewConferenceModal, initStore} from '../../components/conference/NewConference/stores/stores';
+    import {displayModal, initStore} from '../../components/conference/NewConference/stores/stores';
 	import ConferenceCategoryCheckboxGroup from '../../components/checkbox/conference-category-checkbox-group/ConferenceCategoryCheckboxGroup.svelte';
 	import FilterConferenceCategoryModal from '../../components/modal/filter-conference-category-modal/FilterConferenceCategoryModal.svelte';
 	import { goto } from '$app/navigation';
@@ -23,7 +23,7 @@
 	let screenSize: number;
 
 	function openModal() {
-		displayNewConferenceModal.set(true);
+		displayModal.set(true);
 	}
 
 	const categoryParam = urlParams.get('category');
@@ -122,7 +122,7 @@
 </svelte:head>
 
 <div class="conference-page-container">
-	{#if $displayNewConferenceModal}
+	{#if $displayModal}
 		<NewConferenceModal />
 	{/if}
 	<div class="topRow">
